pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "excel-automation"
version = "0.1.0"
description = "Automação de tratamento e padronização de planilhas Excel"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.0",
    "openpyxl>=3.1",
    "xlsxwriter>=3.1",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "typer>=0.9",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
excel-automation = "excel_automation.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=excel_automation --cov-report=term-missing"
